rudimentary object orientation implemented
